Switch time calculations to received time instead of reported time (#7)
This PR changes all time calculations (minutes ago, expiry, etc.) to be based on the time of receipt by the server instead of the reported time in the packet.

This is because some stations are using position reports without timestamps, which breaks both expiry and the "minutes ago" display as implemented.

To this end, a new `WeatherReport<T>` type was implemented to also carry a timestamp. This will allow future expansion as need be. It's generic to allow `WeatherReport<string>` to transmit encoded packets to the client (currently just done as strings) or `WeatherReport<Packet>` to store and manipulate those packets on the client.