Add Connection Retry to AprsIsReceiver (#9)
This change adds connection retry logic to AprsIsReceiver. This is relatively simple logic but should work for the basic case of inactivity disconnection by the APRS-IS server.

While writing this code, I refactored for simplicity, logging, and the ability to stop the server locally more quickly.

Necessary for #6, but perhaps not sufficient as a keep-alive message would allow avoiding disconnection/re-connection flow entirely.