Switch expiry service to Timer (#11)
Currently, the `ReportExpiry` service runs on a single task with a long delay in the middle. This changes that service to use a threading timer instead, which periodically fires a new event on a thread rather than keeping a task running the full time.

This simplifies code and also dramatically speeds up server stop (no timeout is required as it stops successfully now).