Detect User Location (#19)
This change allows using device location to detect gridsquare and fetch a report. Should have basic messaging when unable to fetch location, but nothing too advanced at this point.

Left one oddity in that I can iterate on: looks like nginx casts the request file to lowercase, which was throwing things off. Fixed with a rename in the dockerfile for now, but it'd be ideal to not have to do that in the future.

Validated manually with local docker run.