Only save reports with displayable properties (#25)
This PR updates the logic in the `AprsIsReceiver` class on the server to discard any packets which either:
1. Are not positioned weather reports (aka. `AprsSharp`'s `WeatherInfo` type) (in line with current functionality)
2. Do not carry any of the properties we display in the UX

This will both reduce the amount of reports we have to store on the server AND avoid displaying a report devoid of any information. Interestingly, this appears to result in over 100 fewer packets saved on the server at current time (2,296 instead of 2,409), which is almost 5% of all packets we were storing.