Update Actions for Node versions (#36)
Some actions workflows such as [this one](https://github.com/CBielstein/AprsWeather/actions/runs/4220796087) are reporting warnings for upcoming update to Node runtime version requirements for actions themselves.

This PR bumps up to the highest version of each impacted action to resolve the warning. In this change, we move from setup-dotnet v1 to v3. Although v2 would have sufficed to fix this issue, I'm making the decision to move ahead to the highest version while making a change now to aim to stay up to date and in support.