Reset reportIndex when new location is set. (#37)
This PR fixes a bug where the `reportIndex` was no reset to zero when the `reports` list was reset following a call to `SetLocation` (updating or refreshing location). This means the index could be higher than the count of the reports list and that could result in an exception or odd behavior when:
* the "previous" button is pressed, as the index will decrement and may still be higher than the end of the array when getting the new report (out of bounds exception)
* the "next" button is pushed, potentially causing additional reports to load and the closest few reports to be skipped to catch the list size up to the index

This change simply resets the `reportIndex` to 0 each time the list is reset to ensure none of those scenarios happen.