Move DigitalOcean app spec validation to separate workflow (#44)
Breaks the DIgitalOcean app spec validation out of the main build Action workflow. This allows PRs from forks to run without failing to access secrets. This resolves #43.

I left the validate logic in the deploy step of the build workflow so that the spec is validated again immediately before a deployment is attempted.